Parser API reference

JLPG.JLPG_Parser.Parser — Type

Parser definition.

Created by read_parser_definition_file function.

JLPG.JLPG_Parser.ParserOptions — Type

Parser options definition.

Currently only tag, parser_tag and parser_type options are supported. Option tag defines a tag that will be preprended to all objects generated in the parser file. Option lexer_tag defines a tag that will be preprended to all objects generated associated with the lexer in the parser file. Option parser_type defines the type of the parser to be generated (SLR, LALR or LR).

JLPG.JLPG_Parser.ParserProduction — Type

Parser production definition.

A parser production is a rule of the form:

A -> B1 B2 ... Bn

# or if the production is an alternative to previous production
A -> B1 B2 ... Bn
   | C1 C2 ... Cm

where A is a nonterminal and B1, B2, ..., Bn (C1, C2, ..., Cn) is a string of grammar symbols (terminals or nonterminals).

The action is an optional string of Julia code that is executed when the production is reduced. The action can be used to build an AST or to execute arbitrary code.

The return type is the type of the value returned by the action. If the action is not defined, the return type is nothing.

JLPG.JLPG_Parser.ParsingItem — Type

Parsing item definition.

Used for the construction of the parsing table.

JLPG.JLPG_Parser.ParsingTable — Type

Parsing table definition.

The parsing table is used by the parser to decide what action to take when it is in a certain state and reads a certain symbol.

All parsing table types have the same structure, but they differ in its content.

JLPG.JLPG_Parser.ParsingTableAction — Type

Parsing table action definition.

There are three types of actions:

Shift

Reduce

Accept

The action is executed when the parser is in a certain state and reads a certain symbol.

JLPG.JLPG_Parser.LalrParsingTable — Method

LalrParsingTable(augmented_parser::Parser)

Generate the LALR parsing table for the augmented parser.

The augmented parser is the parser with the augmented start production.

This algorithm computes the LR(1) item sets and gotos, then merges the item sets that are equal. Finally, it generates the parsing table from the merged item sets and gotos (and mappings).

This algorithm is almost the same as the one used for the LR(1) parsing table. The procedure of generating the LR(1) table is described in the Dragon Book, sections 4.7.2, 4.7.3. The idea of merging the item sets is described in the section 4.7.4.

JLPG.JLPG_Parser.LrParsingTable — Method

LrParsingTable(
  augmented_parser::Parser
)::ParsingTable

Compute the LR(1) parsing table for a given parser.

The procedure is described in the Dragon Book, 2nd edition, section 4.7.3.

JLPG.JLPG_Parser.SlrParsingTable — Method

SlrParsingTable(augmented_parser::Parser)

Generate the SLR parsing table for the augmented parser.

The augmented parser is the parser with the augmented start production.

The procedure is described in the Dragon Book, 2nd edition, section 4.6.4.

JLPG.JLPG_Parser.augment_parser — Method

augment_parser(parser::Parser)::Parser

Augment the parser with the augmented start production.

Add the %start nonterminal to the set of nonterminals and the augmented start production to the set of productions.

JLPG.JLPG_Parser.augment_productions — Method

augment_productions(
  starting::Symbol,
  starting_return_type::Symbol,
  productions::Dict{Symbol, Vector{ParserProduction}}
)::Dict{Symbol, Vector{ParserProduction}}

Add the augmented start production to the set of productions.

The augmented start production is of the form:

%start -> starting

where starting is the starting nonterminal of the parser.

JLPG.JLPG_Parser.fill_parser_template — Method

fill_parser_template(
  codeblocks::Vector{String},
  table::ParsingTable,
  productions::Dict{Symbol, Vector{ParserProduction}},
  symbol_types::Dict{Symbol, Symbol}
)::String

Fill the lexer mustache template with the data.

This function is used internally by generate_parser function.

JLPG.JLPG_Parser.first_set_for_string_of_symbols — Method

first_set_for_string_of_symbols(
  symbols::Vector{Symbol},
  firsts::Dict{Symbol, Set{Symbol}}
)::Set{Symbol}

Determine the first set for a string of symbols using precomputed first sets.

JLPG.JLPG_Parser.first_sets — Method

first_sets(
  terminals::Vector{Symbol},
  nonterminals::Vector{Symbol},
  productions::Dict{Symbol, Vector{ParserProduction}}
)::Dict{Symbol, Set{Symbol}}

Compute the first sets for the given grammar.

The first sets are computed using the algorithm provided in the Dragon Book, 2nd edition, pages 220-221.

JLPG.JLPG_Parser.follow_sets — Method

follow_sets(
  firsts::Dict{Symbol, Set{Symbol}},
  terminals::Vector{Symbol},
  nonterminals::Vector{Symbol},
  productions::Dict{Symbol, Vector{ParserProduction}},
  starting::Symbol
)::Dict{Symbol, Set{Symbol}}

Compute the follow sets for the given grammar.

The follow sets are computed using the algorithm provided in the Dragon Book, 2nd edition, pages 221-222.

JLPG.JLPG_Parser.generate_parser — Function

generate_parser(definition_path::String, output_path::String="__PAR__.jl")

Generate a parser source file from a parser definition file.

The parser definition file should follow the syntax described in both read_parser_definition_file and Parser module documentation.

JLPG.JLPG_Parser.lr0_closure — Method

lr0_closure(
  items::Vector{ParsingItem},
  productions::Dict{Symbol, Vector{ParserProduction}},
  nonterminals::Vector{Symbol}
)::Vector{ParsingItem}

Compute the LR(0) closure of a set of items.

The procedure is described in the Dragon Book, 2nd edition, section 4.6.2.

JLPG.JLPG_Parser.lr0_goto — Method

lr0_goto(
  items::Vector{ParsingItem},
  symbol::Symbol,
  productions::Dict{Symbol, Vector{ParserProduction}},
  nonterminals::Vector{Symbol}
)::Vector{ParsingItem}

Compute the LR(0) goto of a set of items.

The procedure is described in the Dragon Book, 2nd edition, section 4.6.2.

JLPG.JLPG_Parser.lr0_items — Method

lr0_items(
  augmented_productions::Dict{Symbol, Vector{ParserProduction}},
  nonterminals::Vector{Symbol},
  grammar_symbols::Vector{Symbol}
)::Tuple{Vector{Vector{ParsingItem}}, Dict{Int, Dict{Symbol, Int}}}

Compute the LR(0) items for the given grammar.

The procedure is described in the Dragon Book, 2nd edition, section 4.6.2.

JLPG.JLPG_Parser.lr1_closure — Method

lr1_closure(
  items::Vector{ParsingItem},
  productions::Dict{Symbol, Vector{ParserProduction}},
  nonterminals::Vector{Symbol},
  firsts::Dict{Symbol, Set{Symbol}},
)::Vector{ParsingItem}

Compute the LR(1) closure of a set of items.

The procedure is described in the Dragon Book, 2nd edition, section 4.7.2.

JLPG.JLPG_Parser.lr1_goto — Method

lr1_goto(
  items::Vector{ParsingItem},
  symbol::Symbol,
  productions::Dict{Symbol, Vector{ParserProduction}},
  nonterminals::Vector{Symbol},
  firsts::Dict{Symbol, Set{Symbol}},
)::Vector{ParsingItem}

Compute the LR(1) goto of a set of items.

The procedure is described in the Dragon Book, 2nd edition, section 4.7.2.

JLPG.JLPG_Parser.lr1_items — Method

lr1_items(
  augmented_productions::Dict{Symbol, Vector{ParserProduction}},
  nonterminals::Vector{Symbol},
  grammar_symbols::Vector{Symbol},
  firsts::Dict{Symbol, Set{Symbol}}
)::Tuple{Vector{Vector{ParsingItem}}, Dict{Int, Dict{Symbol, Int}}}

Compute the LR(1) items and gotos for a given grammar.

The procedure is described in the Dragon Book, 2nd edition, section 4.7.2.

JLPG.JLPG_Parser.lr1_kernels — Method

lr1_kernels(lr1_item_sets::Vector{Vector{ParsingItem}})

Get the kernels of the LR(1) item sets (items without lookahead).

JLPG.JLPG_Parser.merge_lr1_kernels — Method

merge_lr1_kernels(
  lr1_item_sets::Vector{Vector{ParsingItem}},
  lr1_gotos::Dict{Int, Dict{Symbol, Int}},
  lr1_kernels::Vector{Vector{ParsingItem}}
)

Merge the LR(1) kernels that are equal.

Returns a tuple of the merged item sets, gotos and mappings from the old item sets to the new ones.

JLPG.JLPG_Parser.parser_grammar_symbols — Method

parser_grammar_symbols(parser::Parser)

Concatenate parser nonterminals and terminals into a single vector.

JLPG.JLPG_Parser.read_parser_definition_file — Method

read_parser_definition_file(path::String)::Parser

Read parser definition file and construct a Parser object.

The syntax for parser definition files follow the same syntax as GNU Bison definition files.

Definition file structure

Parser definition file consists of three sections:

Definitions section, where lexer tokens, types and parser options are defined.

Productions section, where grammar productions are defined.

Code section, where user code is defined.

Each section is separated by a %% delimiter.

Definitions section

Definitions section consists of lexer tokens, types and parser options.

Lexer tokens

Each lexer token that might be passed to parser during the parsing process must be defined in this section. To define a lexer token, use the following syntax:

%token NAME "ALIAS"

where NAME is the uppercased name of the token and ALIAS is a optional string literal, which may be used to refer to the token in the parser productions section.

Types

Each nonterminal symbol may be assigned a type. To assign a type to a nonterminal, use the following syntax:

%type <type> symbol

where type is the type of the symbol and symbol is the lowercase name of the symbol. The type may be any valid Julia type, including user-defined types.

If a type is not assigned to a nonterminal, it is assumed to be Nothing.

If a type is assigned to a nonterminal, it is assumed that the user will return a value of that type from the action associated with the production. To return a value from the action, assign it to the $$ variable.

Parser options

Parser options may be used to configure the parser. Currently, the following options are available:

tag: Rename the prefix of all objects generated in the parser. This includes all token definitions, special functions, etc.

lexer_tag: Rename the prefix of all objects associated with the lexer that are used in the parser. This includes all token definitions, special functions, etc.

parser_type: Specify the type of the parser. Currently, the following types are available:

SLR

LR

LALR

Parser options are defined as follows:

%option tag="TAG_NAME"
%option lexer_tag="LEXER_TAG_NAME"
%option SLR
%option LR
%option LALR

where TAG_NAME is the new parser prefix, LEXER_TAG_NAME is the new prefix for lexer objects.

Renaming the parser tag will also rename the default name of the generated parser file. For example if the tag is set to MyParser, the generated parser file will be named MyParser.jl.

Productions section

Productions section consists of grammar productions. Each production is defined as follows:

lhs -> production :{ action }:

where lhs is the left-hand side of the production, production is the right-hand side. Right-hand side might consist of terminals (uppercased symbols), nonterminals (lowercased symbols) or an %empty keyword, which represents an empty production. %empty should not be mixed with other symbols in a single production.

The action is an optional Julia code block, which is executed when the production is reduced. The action may return a value, which will be assigned to the left-hand side of the production. To return a value from the action, assign it to the $$ variable:

lhs -> production :{ $$ = value }:

If a given left-hand side has multiple productions, they have to be defined below the first production using the following syntax:

lhs -> production :{ action }:
     | alt_production :{ action }:

To refer to the values of symbols in the right-hand side of the production, use the $<number> syntax, for example:

lhs -> production :{ $$ = $1 + $3 }:

where $n refers to the nth symbol in the right-hand side of the production.

The first production of the first nonterminal is considered as the starting production, unless a special flag is specified. Use %start flag to specify the starting production:

%start lhs

where lhs is the left-hand side of the starting production.

Code section

Code section consists of user code. It is copied to the output file as is, in the same order as it was defined in the definition file.

Additionally the user may define code blocks enclosed with %{ and %}, which will be also copied to the output file. Code blocks may be inserted in all of the sections, but they should not be intermixed with other definition file constructs (such as production definitions, flags, etc.).

Commenting

Comments are defined using the #= and =# delimiters. Comments are single-line only. They should not be mixed with other definition file constructs.

Examples

For valid parser definition files examples see the examples directory.

JLPG.JLPG_Parser.replace_overloaded_functions_in_generated_parser — Method

replace_overloaded_functions_in_generated_parser(
  generated_parser::String
)::String

Replace overloaded functions in the generated parser code.

Scan the generated parser code for overloaded functions and replace the sections between # <<: ovearloaded_func start :>> and # <<: ovearloaded_func end :>> with a message that the function is overloaded.

JLPG.JLPG_Parser.replace_special_tag_in_generated_parser — Method

replace_special_tag_in_generated_parser(
  generated_parser::String,
  parser_tag::String,
  lexer_tag::String
)::String

Replace special lexer and parser tags in the generated parser code.

This function is used to replace the special prefixes present in all generated objects with the user-defined tags.

JLPG.JLPG_Parser.replace_special_variables_in_generated_parser — Method

replace_special_variables_in_generated_parser(
  generated_parser::String
)::String

Replace special variables in the generated parser code.

Currently only special variables are $$ and $n, which are replaced with __PAR__action_result and __PAR__symbols_slice[n] respectively.