SimpleLexer API reference

JLPG.JLPG_SimpleLexer.ArgumentType — Type

All possible types of arguments passed to a lexer token

JLPG.JLPG_SimpleLexer.Lexer — Type

Lexer definition.

Defines a lexer with a set of actions, aliases, code_blocks and options. Created by read_lexer_definition_file function.

JLPG.JLPG_SimpleLexer.LexerAction — Type

Lexer action definition.

Defines a lexer action with a pattern and a body:

pattern :{ body }:

Action body is a string of Julia code that is executed when a pattern is matched.

JLPG.JLPG_SimpleLexer.LexerOptions — Type

Lexer options definition.

Currently only tag option is supported. It defines the tag that will be preprended to all objects generated in the lexer file.

JLPG.JLPG_SimpleLexer.LexerToken — Type

Lexer token type.

By default, all LexerTokens will inherit this type and have some default members generated by the lexer generator.

Example of a generated token

struct Num <: LexerToken
    symbol::Symbol   # This is the name of the token
    values::Dict     # Dict of all values passed during token creation
                     # Those values may be typed, but it is not required
                     # You can access those values by using simple member access notation (token.value)
    file_pos::String # File position of the token
    ???              # Anything else that may be added in the future
end

JLPG.JLPG_SimpleLexer.LexerTokenDefinition — Type

Lexer token definition.

Contains a name and a set of arguments that will be passed to the token constructor inside of generated lexer file.

JLPG.JLPG_SimpleLexer.PatternPart — Type

Possible parts of a pattern in a lexer action.

JLPG.JLPG_SimpleLexer.RegexAlias — Type

Regex alias definition.

Defines a regex alias with a name and a PCRE2-compliant regex pattern:

name pattern

Regex aliases are used to simplify lexer definitions. They are expanded during lexer generation.

Base.getproperty — Method

Base.getproperty(token::LexerToken, name::Symbol)

Support for member access notation (token.value)

JLPG.JLPG_SimpleLexer.expand_regex_aliases_in_actions — Method

expand_regex_aliases_in_actions(
  actions::Vector{LexerAction},
  expanded_aliases::Vector{RegexAlias}
)::Vector{LexerAction}

Expand lexer regex aliases inside of the patterns of actions.

Aliases provided as expanded_aliases should contain previously expanded aliases. Literal strings enclosed in double quotes are escaped using \Q and \E sequences.

JLPG.JLPG_SimpleLexer.expand_regex_aliases_in_aliases — Method

expand_regex_aliases_in_aliases(aliases::Vector{RegexAlias})::Vector{RegexAlias}

Expand lexer regex aliases inside the patterns of other aliases.

Aliases have to be defined before they are referenced.

JLPG.JLPG_SimpleLexer.expand_regex_aliases_in_lexer — Method

expand_regex_aliases_in_lexer(lexer::Lexer)::Lexer

Expand aliases and actions in lexer into proper regexes.

JLPG.JLPG_SimpleLexer.fill_lexer_template — Method

fill_lexer_template(
  tokens::Vector{LexerTokenDefinition},
  codeblocks::Vector{String},
  actions::Vector{LexerAction}
)::String

Fill the lexer mustache template with the given tokens, codeblocks and actions.

This function is used internally by generate_lexer function.

JLPG.JLPG_SimpleLexer.generate_lexer — Function

generate_lexer(definition_path::String, output_path::String="__LEX__.jl")

Generate a lexer source file from a lexer definition file.

The lexer definition file should follow the syntax described in both read_lexer_definition_file and SimpleLexer module documentation.

By default parsers assume that there is a lexer file named __LEX__.jl in the same directory as the parser file. The user may specify a different path by passing output_path argument. If another path is specified, it is required that the user manually include the generated lexer file in the parser definition file.

JLPG.JLPG_SimpleLexer.read_lexer_definition_file — Method

read_lexer_definition_file(path::String)::Lexer

Read lexer definition file from path and construct a Lexer object.

The syntax for lexer definition files follow the same syntax as Flex definition files.

Definition file structure

Lexer definition file consists of three sections:

Definitions section, where regex aliases and lexer options are defined.

Actions section, where lexer actions are defined.

Code section, where user code is defined.

Each section is separated by a %% delimiter.

Definitions section

Definitions section consists of regex aliases and lexer options. Regexes should conform to the PCRE syntax, since Julia uses PCRE regexes.

Regex aliases

Regex aliases are defined as follows:

ALIAS_NAME REGEX

where ALIAS_NAME is a name of the alias and REGEX is a regex pattern.

Regex aliases can be referenced in other regexes by using {ALIAS_NAME} syntax. If an alias is referenced before it is defined, an error will be thrown.

Lexer options

Currently there's only one lexer option available: tag. This option allows renaming of the prefix of all objects generated in the lexer. This includes all token definitions, special functions, etc.

The tag option is defined as follows:

%option tag="TAG_NAME"

where TAG_NAME is the new prefix.

Renaming the lexer tag will also rename the default name of the generated lexer file. For example, if the tag is set to MyLexer, the generated lexer file will be named MyLexer.jl. Keep in mind, that a generated parser requires __LEX__.jl file to be present in the same directory. If no such file is found, it is required that the user includes it by themselves using in a code block/section.

Actions section

Actions section consists of lexer actions. Each lexer action is defined as follows:

PATTERN :{ ACTION_BODY }:

where PATTERN is a regex pattern and ACTION_BODY is a Julia code block.

The PATTERN is matched against the input text. If the match is successful, the ACTION_BODY is executed. The ACTION_BODY should return a token object, but it is not required.

Action patterns

Action patterns are constructed using one or many pattern parts: regexes, aliases or literal strings. Patterns should not be separated by whitespace, since it is treated as a literal whitespace character. Literal strings are strings enclosed in double quotes ("). They are treated as a sequence of characters, not as a regex. Regexes are defined using the PCRE syntax.

For example:

"hello"{ALIAS_NAME}[0-9]+

is a valid pattern definition, which will be later converted to a single regex.

Action tokens

Each action body may return a token object using the return keyword. There is no need to predefine token types, since they are generated automatically by retrieve_tokens_from_actions function.

Tokens are defined as follows:

return TOKEN_NAME(ARGUMENTS)

where TOKEN_NAME is a name of the token and ARGUMENTS is a list of arguments. Arguments may be named and typed, but it is not required. See: retrieve_tokens_from_actions for more information.

Code section

Code section consists of user code. It is copied to the output file as is, in the same order as it was defined in the definition file.

Additionally the user may define code blocks enclosed with %{ and %}, which will be also copied to the output file. Code blocks may be inserted in all of the sections, but they should not be intermixed with other definition file constructs (such as action patterns, action bodies, etc.).

Commenting

Comments are defined using the #= and =# delimiters. Comments are single-line only. They should not be mixed with other definition file constructs.

Example

For valid lexer definition files examples see the examples directory.

JLPG.JLPG_SimpleLexer.replace_overloaded_functions_in_generated_lexer — Method

replace_overloaded_functions_in_generated_lexer(
  generated_lexer::String
)::String

Replace overloaded functions in the generated lexer code.

Scan the generated lexer code for overloaded functions and replace the sections between # <<: ovearloaded_func start :>> and # <<: ovearloaded_func end :>> with a message that the function is overloaded.

JLPG.JLPG_SimpleLexer.replace_special_tag_in_generated_lexer — Method

replace_special_tag_in_generated_lexer(
  generated_lexer::String,
  tag::String
)::String

Replace special tag in the generated lexer code.

This function is used to replace the special prefix present in all generated objects with the user-defined tag.

JLPG.JLPG_SimpleLexer.replace_special_variables_in_generated_lexer — Method

replace_special_variables_in_generated_lexer(
  generated_lexer::String
)::String

Replace special variables in the generated lexer code.

Currently the only special variable is $$, which is replaced with __LEX__current_match().

JLPG.JLPG_SimpleLexer.replace_token_args_in_actions — Method

replace_token_args_in_actions(
  lexer::Lexer,
  defined_tokens::Vector{LexerTokenDefinition}
)::Lexer

Convert token definitions from lexer definition file to conform to the code generation template.

This function does two things:

Prepends a special lexer tag to each token name. This is done to avoid name collisions with user-defined functions.

Replaces arguments passed to returned tokens with a list of keyword arguments. Argument names should be defined in the defined_tokens vector.

For example return Number(1) will become return __LEX__Number(;value=1).

JLPG.JLPG_SimpleLexer.retrieve_tokens_from_actions — Method

retrieve_tokens_from_actions(actions::Vector{LexerAction})::Vector{LexerTokenDefinition}

Scan lexer actions for returned tokens and their arguments.

Each lexer action may return some sort of token, which is later passed to the parser. Tokens may contain additional parameters of any type interpretable by Julia. All parameters may be named and typed, but it is not required.

Argument naming rules:

If a token has only one parameter and it is not named, it will be named as value.

If a token has more than one parameter and they are not named, they will be named as value1, value2, etc.

If a custom name is specified, it will be used instead of the default name.

Argument typing rules:

If no type is specified, the value will be a String.

Otherwise, the value will be treated as a value of the specified type. The user should make sure that a proper value for this type is passed.

Examples:

{NUM} { return Num(5) } A type for the first argument is not specified, so it will be treated as a String

{NUM} { return Num(::Int=5) } Num has value of type Int

{ID} { return ID("hello", "world", ::Int=4)} ID has 3 arguments, all of which will be retrieveable by using token.value1, token.value2, token.value3

{ID} { return ID(first::String="hello", second::String="world", num::Int=4)} ID has 3 arguments, all of which will be retrieveable by using token.first, token.second, token.num

JLPG.JLPG_SimpleLexer.token_file_pos — Method

token_file_pos(token::LexerToken)::String

Return the file position of the token.

JLPG.JLPG_SimpleLexer.token_symbol — Method

token_symbol(token::LexerToken)::Symbol

Return the symbol associated with the token.

JLPG.JLPG_SimpleLexer.token_values — Method

token_values(token::LexerToken)::Dict

Return all values passed during token creation.